[ADD] blockchain: The basic outline and functionality of a 'Block' and 'BlockChain'
The 'BlockChain' contains a list of 'Block's and contains functionality to add a block
to the 'chain' or list. The 'Block' currently calculates the hash and stores the data
and previous hash of the previous block. __str__ is overloaded to enable a terminal
graphical output of the BlockChain when printed.